Axing a C5.0.

C5.0 objects are created from the C50 package, which provides an interface to the C5.0 classification model. The models that can be generated include basic tree-based models as well as rule-based models.

# S3 method for C5.0
axe_call(x, verbose = FALSE, ...)

# S3 method for C5.0
axe_ctrl(x, verbose = FALSE, ...)

# S3 method for C5.0
axe_fitted(x, verbose = FALSE, ...)

Arguments

x
A model object.

verbose
Print information each time an axe method is executed. Notes how much memory is released and what functions are disabled. Default is FALSE.

...
Any additional arguments related to axing.

Value

Axed C5.0 object.

Examples

# Load libraries
suppressWarnings(suppressMessages(library(parsnip)))
suppressWarnings(suppressMessages(library(rsample)))
suppressWarnings(suppressMessages(library(rpart)))

# Load data
set.seed(1234)
split <- initial_split(kyphosis, props = 9/10)
spine_train <- training(split)

# Create model and fit
c5_fit <- decision_tree(mode = "classification") %>%
  set_engine("C5.0") %>%
  fit(Kyphosis ~ ., data = spine_train)

out <- butcher(c5_fit, verbose = TRUE)

#> ✔ Memory released: '26,752 B'
#> ✖ Disabled: `print()`, `summary()`, `C5.0Control()`, `C5imp()`

# Try another model from parsnip
c5_fit2 <- boost_tree(mode = "classification", trees = 100) %>%
  set_engine("C5.0") %>%
  fit(Kyphosis ~ ., data = spine_train)
out <- butcher(c5_fit2, verbose = TRUE)

#> ✔ Memory released: '43,368 B'
#> ✖ Disabled: `print()`, `summary()`, `C5.0Control()`, `C5imp()`

# Create model object from original library
library(C50)
data(churn)
c5_fit3 <- C5.0(x = churnTrain[, -20], y = churnTrain$churn)
out <- butcher(c5_fit3, verbose = TRUE)

#> ✔ Memory released: '5,560 B'
#> ✖ Disabled: `print()`, `summary()`, `C5.0Control()`, `C5imp()`