Axing a classbagg object.

classbagg objects are created from the ipred package, which leverages various resampling and bagging techniques to improve predictive models.

# S3 method for classbagg
axe_call(x, verbose = FALSE, ...)

# S3 method for classbagg
axe_data(x, verbose = FALSE, ...)

# S3 method for classbagg
axe_env(x, verbose = FALSE, ...)

Arguments

x
A model object.

verbose
Print information each time an axe method is executed. Notes how much memory is released and what functions are disabled. Default is FALSE.

...
Any additional arguments related to axing.

Value

Axed classbagg object.

Examples

# Load libraries
suppressWarnings(suppressMessages(library(ipred)))
suppressWarnings(suppressMessages(library(rpart)))
suppressWarnings(suppressMessages(library(MASS)))

# Load data
data("GlaucomaM", package = "TH.data")

classbagg_fit <- bagging(Class ~ ., data = GlaucomaM, coob = TRUE)

butcher(classbagg_fit)

#> 
#> Bagging classification trees with 25 bootstrap replications 
#> 
#> Call: dummy_call()
#> 
#> Out-of-bag estimate of misclassification error:  0.1786 
#>