Axing an cv.glmnet.

This is where all the cv.glmnet specific documentation lies.

# S3 method for cv.glmnet
axe_fitted(x, verbose = FALSE, ...)

Arguments

x
Model object.

verbose
Print information each time an axe method is executed that notes how much memory is released and what functions are disabled. Default is FALSE.

...
Any additional arguments related to axing.

Value

Axed model object.

Examples

# Load libraries
library(glmnet)

#> Loading required package: Matrix

#> 
#> Attaching package: ‘Matrix’

#> The following object is masked from ‘package:tidyr’:
#> 
#>     expand

#> Loading required package: foreach

#> Loaded glmnet 2.0-18

n <- 500;p <- 30
nzc <- trunc(p/10)
x <- matrix(rnorm(n*p), n, p)
beta3 <- matrix(rnorm(30), 10, nzc)
beta3 <- rbind(beta3, matrix(0, p-10, nzc))
f3 <- x %*% beta3
p3 <- exp(f3)
p3 <- p3/apply(p3, 1, sum)
g3 <- rmult(p3)
set.seed(10101)
cvfit <- cv.glmnet(x, g3, family="multinomial", keep = TRUE)

butcher(cvfit)

#> $lambda
#>  [1] 0.2137442207 0.1947557732 0.1774542071 0.1616896645 0.1473256003
#>  [6] 0.1342376000 0.1223123016 0.1114464139 0.1015458216 0.0925247707
#> [11] 0.0843051251 0.0768156902 0.0699915960 0.0637737355 0.0581082526
#> [16] 0.0529460756 0.0482424922 0.0439567621 0.0400517645 0.0364936762
#> [21] 0.0332516786 0.0302976911 0.0276061277 0.0251536754 0.0229190922
#> [26] 0.0208830232 0.0190278330 0.0173374528 0.0157972413 0.0143938579
#> [31] 0.0131151472 0.0119500337 0.0108884256 0.0099211279 0.0090397622
#> [36] 0.0082366947 0.0075049694 0.0068382486 0.0062307575 0.0056772342
#> [41] 0.0051728843 0.0047133395 0.0042946194 0.0039130973 0.0035654685
#> [46] 0.0032487221 0.0029601146 0.0026971462 0.0024575391 0.0022392181
#> [51] 0.0020402922 0.0018590382 0.0016938864 0.0015434062 0.0014062943
#> [56] 0.0012813629 0.0011675302 0.0010638100 0.0009693041 0.0008831938
#> [61] 0.0008047333 0.0007332430 0.0006681037 0.0006087512 0.0005546714
#> [66] 0.0005053959 0.0004604980 0.0004195886 0.0003823135 0.0003483498
#> [71] 0.0003174034 0.0002892061 0.0002635139 0.0002401040 0.0002187738
#> [76] 0.0001993386
#> 
#> $cvm
#>  [1] 2.204299 2.165054 2.099082 2.029062 1.967170 1.909373 1.855444 1.798674
#>  [9] 1.740753 1.686552 1.631171 1.572995 1.518931 1.469644 1.423935 1.380970
#> [17] 1.340535 1.303650 1.270114 1.239095 1.210102 1.184239 1.161574 1.141882
#> [25] 1.124620 1.109995 1.097595 1.087352 1.078485 1.071383 1.065856 1.061480
#> [33] 1.058056 1.055678 1.054064 1.053420 1.053688 1.054929 1.057193 1.060108
#> [41] 1.063265 1.066716 1.070398 1.074169 1.078033 1.081870 1.085778 1.089813
#> [49] 1.093825 1.097765 1.101664 1.105451 1.109132 1.112722 1.116218 1.119536
#> [57] 1.122712 1.125710 1.128553 1.131232 1.133745 1.136091 1.138273 1.140318
#> [65] 1.142257 1.144055 1.145702 1.147245 1.148640 1.149904 1.151073 1.151893
#> [73] 1.152526 1.153062 1.153512 1.153854
#> 
#> $cvsd
#>  [1] 0.004733315 0.004535849 0.006139857 0.008540510 0.011135881 0.013369859
#>  [7] 0.015537732 0.016892797 0.017787047 0.018756283 0.020073514 0.020969090
#> [13] 0.021965983 0.022932757 0.023761887 0.024738078 0.025556433 0.026387263
#> [19] 0.027286428 0.028071434 0.028720353 0.029434633 0.029933451 0.030324483
#> [25] 0.030729184 0.031208243 0.031724433 0.032262682 0.032608148 0.032898232
#> [31] 0.033141614 0.033427198 0.033713871 0.033919080 0.034088028 0.034447877
#> [37] 0.034842517 0.035248747 0.035655899 0.036051365 0.036466426 0.036952332
#> [43] 0.037483707 0.038056297 0.038631649 0.039211981 0.039782568 0.040346504
#> [49] 0.040916520 0.041418948 0.041904073 0.042374032 0.042814168 0.043219215
#> [55] 0.043609520 0.043963362 0.044316090 0.044667514 0.044999660 0.045306757
#> [61] 0.045593972 0.045858313 0.046097731 0.046324928 0.046550852 0.046765804
#> [67] 0.046964588 0.047144890 0.047316793 0.047463009 0.047620762 0.047680342
#> [73] 0.047708080 0.047723303 0.047741759 0.047777088
#> 
#> $cvup
#>  [1] 2.209032 2.169589 2.105222 2.037602 1.978306 1.922743 1.870981 1.815567
#>  [9] 1.758540 1.705308 1.651245 1.593964 1.540897 1.492577 1.447696 1.405708
#> [17] 1.366092 1.330037 1.297400 1.267166 1.238822 1.213673 1.191507 1.172206
#> [25] 1.155349 1.141203 1.129320 1.119615 1.111093 1.104282 1.098998 1.094907
#> [33] 1.091769 1.089597 1.088152 1.087867 1.088531 1.090178 1.092848 1.096159
#> [41] 1.099731 1.103668 1.107882 1.112225 1.116664 1.121082 1.125561 1.130160
#> [49] 1.134741 1.139184 1.143568 1.147825 1.151946 1.155942 1.159827 1.163500
#> [57] 1.167029 1.170377 1.173553 1.176539 1.179339 1.181950 1.184371 1.186643
#> [65] 1.188807 1.190821 1.192667 1.194389 1.195956 1.197367 1.198694 1.199574
#> [73] 1.200234 1.200785 1.201253 1.201631
#> 
#> $cvlo
#>  [1] 2.199565 2.160518 2.092942 2.020521 1.956034 1.896003 1.839906 1.781781
#>  [9] 1.722966 1.667795 1.611098 1.552026 1.496965 1.446711 1.400173 1.356232
#> [17] 1.314979 1.277263 1.242827 1.211023 1.181381 1.154804 1.131640 1.111557
#> [25] 1.093891 1.078787 1.065871 1.055090 1.045877 1.038485 1.032715 1.028053
#> [33] 1.024342 1.021759 1.019976 1.018972 1.018846 1.019680 1.021537 1.024057
#> [41] 1.026798 1.029764 1.032915 1.036113 1.039401 1.042658 1.045996 1.049467
#> [49] 1.052908 1.056346 1.059760 1.063077 1.066318 1.069503 1.072608 1.075573
#> [57] 1.078396 1.081042 1.083554 1.085926 1.088151 1.090233 1.092175 1.093993
#> [65] 1.095706 1.097289 1.098738 1.100100 1.101323 1.102441 1.103452 1.104213
#> [73] 1.104818 1.105339 1.105770 1.106077
#> 
#> $nzero
#>  s0  s1  s2  s3  s4  s5  s6  s7  s8  s9 s10 s11 s12 s13 s14 s15 s16 s17 s18 s19 
#>   0   1   1   1   1   2   2   3   3   3   3   3   4   4   6   6   6   6   6   6 
#> s20 s21 s22 s23 s24 s25 s26 s27 s28 s29 s30 s31 s32 s33 s34 s35 s36 s37 s38 s39 
#>   6   7   7   9   9  10  10  10  11  11  12  14  15  15  15  17  17  17  17  17 
#> s40 s41 s42 s43 s44 s45 s46 s47 s48 s49 s50 s51 s52 s53 s54 s55 s56 s57 s58 s59 
#>  17  17  19  19  19  19  19  19  19  19  19  19  19  19  19  19  19  19  19  19 
#> s60 s61 s62 s63 s64 s65 s66 s67 s68 s69 s70 s71 s72 s73 s74 s75 
#>  19  19  19  19  20  20  20  20  20  20  20  20  20  20  20  20 
#> 
#> $name
#>               deviance 
#> "Multinomial Deviance" 
#> 
#> $glmnet.fit
#> 
#> Call:  glmnet(x = x, y = g3, family = "multinomial") 
#> 
#>       Df      %Dev    Lambda
#>  [1,]  0 4.650e-15 0.2137000
#>  [2,]  2 1.901e-02 0.1948000
#>  [3,]  3 5.110e-02 0.1775000
#>  [4,]  3 8.311e-02 0.1617000
#>  [5,]  3 1.117e-01 0.1473000
#>  [6,]  3 1.391e-01 0.1342000
#>  [7,]  3 1.634e-01 0.1223000
#>  [8,]  5 1.921e-01 0.1114000
#>  [9,]  5 2.189e-01 0.1015000
#> [10,]  7 2.434e-01 0.0925200
#> [11,]  8 2.725e-01 0.0843100
#> [12,]  8 3.010e-01 0.0768200
#> [13,]  9 3.265e-01 0.0699900
#> [14,]  9 3.509e-01 0.0637700
#> [15,]  9 3.738e-01 0.0581100
#> [16,]  9 3.954e-01 0.0529500
#> [17,]  9 4.148e-01 0.0482400
#> [18,]  9 4.322e-01 0.0439600
#> [19,] 11 4.488e-01 0.0400500
#> [20,] 11 4.645e-01 0.0364900
#> [21,] 11 4.785e-01 0.0332500
#> [22,] 12 4.914e-01 0.0303000
#> [23,] 13 5.034e-01 0.0276100
#> [24,] 15 5.147e-01 0.0251500
#> [25,] 16 5.252e-01 0.0229200
#> [26,] 19 5.350e-01 0.0208800
#> [27,] 20 5.442e-01 0.0190300
#> [28,] 20 5.525e-01 0.0173400
#> [29,] 21 5.603e-01 0.0158000
#> [30,] 21 5.675e-01 0.0143900
#> [31,] 22 5.740e-01 0.0131200
#> [32,] 23 5.799e-01 0.0119500
#> [33,] 25 5.856e-01 0.0108900
#> [34,] 25 5.908e-01 0.0099210
#> [35,] 25 5.954e-01 0.0090400
#> [36,] 25 5.996e-01 0.0082370
#> [37,] 25 6.033e-01 0.0075050
#> [38,] 27 6.066e-01 0.0068380
#> [39,] 28 6.096e-01 0.0062310
#> [40,] 29 6.123e-01 0.0056770
#> [41,] 29 6.147e-01 0.0051730
#> [42,] 29 6.169e-01 0.0047130
#> [43,] 29 6.188e-01 0.0042950
#> [44,] 29 6.204e-01 0.0039130
#> [45,] 30 6.219e-01 0.0035650
#> [46,] 30 6.232e-01 0.0032490
#> [47,] 30 6.243e-01 0.0029600
#> [48,] 30 6.253e-01 0.0026970
#> [49,] 30 6.261e-01 0.0024580
#> [50,] 30 6.269e-01 0.0022390
#> [51,] 30 6.275e-01 0.0020400
#> [52,] 30 6.280e-01 0.0018590
#> [53,] 30 6.285e-01 0.0016940
#> [54,] 30 6.289e-01 0.0015430
#> [55,] 30 6.292e-01 0.0014060
#> [56,] 30 6.295e-01 0.0012810
#> [57,] 30 6.297e-01 0.0011680
#> [58,] 30 6.299e-01 0.0010640
#> [59,] 30 6.301e-01 0.0009693
#> [60,] 30 6.302e-01 0.0008832
#> [61,] 30 6.304e-01 0.0008047
#> [62,] 30 6.305e-01 0.0007332
#> [63,] 30 6.306e-01 0.0006681
#> [64,] 30 6.306e-01 0.0006088
#> [65,] 30 6.307e-01 0.0005547
#> [66,] 30 6.307e-01 0.0005054
#> [67,] 30 6.308e-01 0.0004605
#> [68,] 30 6.308e-01 0.0004196
#> [69,] 30 6.308e-01 0.0003823
#> [70,] 30 6.309e-01 0.0003483
#> [71,] 30 6.309e-01 0.0003174
#> [72,] 30 6.309e-01 0.0002892
#> [73,] 30 6.309e-01 0.0002635
#> [74,] 30 6.309e-01 0.0002401
#> [75,] 30 6.310e-01 0.0002188
#> [76,] 30 6.310e-01 0.0001993
#> 
#> $fit.preval
#> numeric(0)
#> 
#> $foldid
#>   [1]  4  1  6  7 10  2 10  4  5  5  4  1  3  8  5  3 10  9  3  3  8  7  7  9  2
#>  [26]  5  6  2  8  4  3  8  3  1  9  4  2  6  2  5  8  1  5  7  1  2  7  1  6  8
#>  [51]  7  5  3  4 10  6  2  7  2  4  6 10  6  8  2  3  2  3  4  2  6  7  7  9  4
#>  [76]  2 10  8  6  6  2  2  2  4 10  4  9  6  5  3  9 10  9  8  1  7  9  4  5  5
#> [101]  8  4  2 10  5  9  9  9  4  2  5  3  6  6 10  7  2  9  2  1 10 10  6  6  3
#> [126]  5  5  4  9  9  1  9 10  2  8  4  2  8  5  9  3  4  6  9  8  7  7 10  3  7
#> [151]  6  1  7  1  9  1 10  4  7  3 10  9  4  8  9  5  9  8  5  7  7  7  3  1  3
#> [176]  5  2 10  2  4  8  6  4  8  5 10  6  4  8  1  9  6  5  4 10 10  9  4  3  2
#> [201]  8 10  4 10 10  4  8  3  2  2  3  4  8  7  1  8  9  6  3  4  7  2  7  7  8
#> [226]  1  8  1  7  9  6  4  7 10  2  6 10  4  2  1 10  7  3 10  6  8  7  2  6  3
#> [251]  1 10 10  3  6  2  4  7  9  2  4  3  9  7  5  9 10  4  7  2  5  5  6  1  7
#> [276]  2  1  5  3  3  1  8  6  4  3  8 10  4  5  5  8  2  7  3  2  6  1  5 10  1
#> [301]  7  1  3  5  9  9  5  3  5  6  2  2  3  9  1  9 10  9  5  9  6 10 10  8  8
#> [326]  5  5  6  3  6  7  6  7  9  4  3  9  2  9  3  9  3  7  3  9  6  2  7  6  8
#> [351]  8  2  9  5 10  4  2  1  6  8 10  6  1  1  7  5  7  3  9  6  3  1 10  3  7
#> [376]  8  1  5  5  6  4  5  5  3  7  8  5  6  3  1  4  2  6  3 10  9  8  3 10  1
#> [401]  7  6  8  8  1  2  9  1  6  2  7  4  3  4  1  5  5  4  8  1  1  1  8  7  1
#> [426]  1  6  9  9  3  5 10  6  8  2  3  2  8  6  3  2  2  9  7  5  8  4  1  8  7
#> [451]  7 10  6  9  5  1  7  4  5  7 10  7  8  3 10 10  8  1 10  1 10  9  2  8  4
#> [476]  8  5 10  4  3  9  1  1  8  5  5  9  1 10  6  1  5  4  1 10  6  4  4  8  4
#> 
#> $lambda.min
#> [1] 0.008236695
#> 
#> $lambda.1se
#> [1] 0.01733745
#> 
#> attr(,"class")
#> [1] "butchered_cv.glmnet" "cv.glmnet"