Axing an elnet.

elnet objects are created from the glmnet package, leveraged to fit generalized linear models via penalized maximum likelihood.

# S3 method for elnet
axe_call(x, verbose = FALSE, ...)

Arguments

x
A model object.

verbose
Print information each time an axe method is executed. Notes how much memory is released and what functions are disabled. Default is FALSE.

...
Any additional arguments related to axing.

Value

Axed model object.

Examples

# Load libraries
suppressWarnings(suppressMessages(library(parsnip)))
suppressWarnings(suppressMessages(library(rsample)))

# Load data
split <- initial_split(mtcars, props = 9/10)
car_train <- training(split)
car_test  <- testing(split)

# Create model and fit
elnet_fit <- linear_reg(mixture = 0, penalty = 0.1) %>%
  set_engine("glmnet") %>%
  fit_xy(x = car_train[, 2:11], y = car_train[, 1, drop = FALSE])

butcher(elnet_fit)

#> 
#> Call:  dummy_call() 
#> 
#>        Df      %Dev    Lambda
#>   [1,] 10 9.494e-36 5132.0000
#>   [2,] 10 1.184e-02 4676.0000
#>   [3,] 10 1.298e-02 4260.0000
#>   [4,] 10 1.423e-02 3882.0000
#>   [5,] 10 1.559e-02 3537.0000
#>   [6,] 10 1.709e-02 3223.0000
#>   [7,] 10 1.872e-02 2937.0000
#>   [8,] 10 2.052e-02 2676.0000
#>   [9,] 10 2.247e-02 2438.0000
#>  [10,] 10 2.461e-02 2221.0000
#>  [11,] 10 2.695e-02 2024.0000
#>  [12,] 10 2.951e-02 1844.0000
#>  [13,] 10 3.230e-02 1680.0000
#>  [14,] 10 3.534e-02 1531.0000
#>  [15,] 10 3.867e-02 1395.0000
#>  [16,] 10 4.229e-02 1271.0000
#>  [17,] 10 4.623e-02 1158.0000
#>  [18,] 10 5.052e-02 1055.0000
#>  [19,] 10 5.519e-02  961.6000
#>  [20,] 10 6.027e-02  876.2000
#>  [21,] 10 6.578e-02  798.3000
#>  [22,] 10 7.176e-02  727.4000
#>  [23,] 10 7.824e-02  662.8000
#>  [24,] 10 8.526e-02  603.9000
#>  [25,] 10 9.284e-02  550.3000
#>  [26,] 10 1.010e-01  501.4000
#>  [27,] 10 1.098e-01  456.8000
#>  [28,] 10 1.193e-01  416.3000
#>  [29,] 10 1.295e-01  379.3000
#>  [30,] 10 1.405e-01  345.6000
#>  [31,] 10 1.522e-01  314.9000
#>  [32,] 10 1.647e-01  286.9000
#>  [33,] 10 1.780e-01  261.4000
#>  [34,] 10 1.921e-01  238.2000
#>  [35,] 10 2.070e-01  217.0000
#>  [36,] 10 2.228e-01  197.8000
#>  [37,] 10 2.395e-01  180.2000
#>  [38,] 10 2.570e-01  164.2000
#>  [39,] 10 2.752e-01  149.6000
#>  [40,] 10 2.942e-01  136.3000
#>  [41,] 10 3.140e-01  124.2000
#>  [42,] 10 3.343e-01  113.2000
#>  [43,] 10 3.553e-01  103.1000
#>  [44,] 10 3.767e-01   93.9500
#>  [45,] 10 3.986e-01   85.6000
#>  [46,] 10 4.207e-01   78.0000
#>  [47,] 10 4.431e-01   71.0700
#>  [48,] 10 4.654e-01   64.7600
#>  [49,] 10 4.877e-01   59.0000
#>  [50,] 10 5.099e-01   53.7600
#>  [51,] 10 5.316e-01   48.9900
#>  [52,] 10 5.530e-01   44.6300
#>  [53,] 10 5.738e-01   40.6700
#>  [54,] 10 5.939e-01   37.0600
#>  [55,] 10 6.133e-01   33.7600
#>  [56,] 10 6.318e-01   30.7600
#>  [57,] 10 6.494e-01   28.0300
#>  [58,] 10 6.661e-01   25.5400
#>  [59,] 10 6.818e-01   23.2700
#>  [60,] 10 6.965e-01   21.2000
#>  [61,] 10 7.102e-01   19.3200
#>  [62,] 10 7.228e-01   17.6000
#>  [63,] 10 7.345e-01   16.0400
#>  [64,] 10 7.453e-01   14.6200
#>  [65,] 10 7.552e-01   13.3200
#>  [66,] 10 7.642e-01   12.1300
#>  [67,] 10 7.725e-01   11.0600
#>  [68,] 10 7.800e-01   10.0700
#>  [69,] 10 7.868e-01    9.1790
#>  [70,] 10 7.930e-01    8.3630
#>  [71,] 10 7.987e-01    7.6200
#>  [72,] 10 8.038e-01    6.9440
#>  [73,] 10 8.085e-01    6.3270
#>  [74,] 10 8.127e-01    5.7650
#>  [75,] 10 8.166e-01    5.2530
#>  [76,] 10 8.202e-01    4.7860
#>  [77,] 10 8.234e-01    4.3610
#>  [78,] 10 8.265e-01    3.9730
#>  [79,] 10 8.292e-01    3.6200
#>  [80,] 10 8.318e-01    3.2990
#>  [81,] 10 8.342e-01    3.0060
#>  [82,] 10 8.364e-01    2.7390
#>  [83,] 10 8.386e-01    2.4950
#>  [84,] 10 8.405e-01    2.2740
#>  [85,] 10 8.424e-01    2.0720
#>  [86,] 10 8.442e-01    1.8880
#>  [87,] 10 8.459e-01    1.7200
#>  [88,] 10 8.476e-01    1.5670
#>  [89,] 10 8.492e-01    1.4280
#>  [90,] 10 8.508e-01    1.3010
#>  [91,] 10 8.523e-01    1.1860
#>  [92,] 10 8.538e-01    1.0800
#>  [93,] 10 8.553e-01    0.9842
#>  [94,] 10 8.567e-01    0.8968
#>  [95,] 10 8.581e-01    0.8171
#>  [96,] 10 8.595e-01    0.7445
#>  [97,] 10 8.609e-01    0.6784
#>  [98,] 10 8.622e-01    0.6181
#>  [99,] 10 8.636e-01    0.5632
#> [100,] 10 8.648e-01    0.5132