Axing an multnet.

multnet objects are created from carrying out multinomial regression in the glmnet package.

# S3 method for multnet
axe_call(x, verbose = FALSE, ...)

Arguments

x
A model object.

verbose
Print information each time an axe method is executed. Notes how much memory is released and what functions are disabled. Default is FALSE.

...
Any additional arguments related to axing.

Value

Axed multnet object.

Examples

# Load libraries
suppressWarnings(suppressMessages(library(parsnip)))

# Load data
set.seed(1234)
predictrs <- matrix(rnorm(100*20), ncol = 20)
response <- as.factor(sample(1:4, 100, replace = TRUE))

# Create model and fi
multnet_fit <- multinom_reg() %>%
  set_engine("glmnet") %>%
  fit_xy(x = predictrs, y = response)

butcher(multnet_fit)

#> 
#> Call:  dummy_call() 
#> 
#>       Df       %Dev    Lambda
#>  [1,]  0 -1.445e-15 0.1010000
#>  [2,]  2  5.267e-03 0.0919900
#>  [3,]  2  1.092e-02 0.0838200
#>  [4,]  2  1.560e-02 0.0763700
#>  [5,]  3  1.963e-02 0.0695900
#>  [6,]  5  2.632e-02 0.0634100
#>  [7,]  5  3.377e-02 0.0577700
#>  [8,]  9  4.214e-02 0.0526400
#>  [9,] 11  5.270e-02 0.0479600
#> [10,] 13  6.365e-02 0.0437000
#> [11,] 15  7.531e-02 0.0398200
#> [12,] 17  8.862e-02 0.0362800
#> [13,] 17  1.007e-01 0.0330600
#> [14,] 17  1.115e-01 0.0301200
#> [15,] 18  1.212e-01 0.0274500
#> [16,] 19  1.300e-01 0.0250100
#> [17,] 19  1.381e-01 0.0227900
#> [18,] 19  1.452e-01 0.0207600
#> [19,] 19  1.517e-01 0.0189200
#> [20,] 20  1.574e-01 0.0172400
#> [21,] 20  1.626e-01 0.0157100
#> [22,] 20  1.669e-01 0.0143100
#> [23,] 20  1.707e-01 0.0130400
#> [24,] 20  1.744e-01 0.0118800
#> [25,] 20  1.776e-01 0.0108300
#> [26,] 20  1.804e-01 0.0098640
#> [27,] 20  1.829e-01 0.0089880
#> [28,] 20  1.850e-01 0.0081890
#> [29,] 20  1.869e-01 0.0074620
#> [30,] 20  1.885e-01 0.0067990
#> [31,] 20  1.898e-01 0.0061950
#> [32,] 20  1.909e-01 0.0056450
#> [33,] 20  1.919e-01 0.0051430
#> [34,] 20  1.928e-01 0.0046860
#> [35,] 20  1.936e-01 0.0042700
#> [36,] 20  1.943e-01 0.0038910
#> [37,] 20  1.949e-01 0.0035450
#> [38,] 20  1.954e-01 0.0032300
#> [39,] 20  1.958e-01 0.0029430
#> [40,] 20  1.962e-01 0.0026820
#> [41,] 20  1.965e-01 0.0024430
#> [42,] 20  1.967e-01 0.0022260
#> [43,] 20  1.970e-01 0.0020290
#> [44,] 20  1.971e-01 0.0018480
#> [45,] 20  1.973e-01 0.0016840
#> [46,] 20  1.974e-01 0.0015350
#> [47,] 20  1.975e-01 0.0013980
#> [48,] 20  1.976e-01 0.0012740
#> [49,] 20  1.977e-01 0.0011610
#> [50,] 20  1.978e-01 0.0010580
#> [51,] 20  1.978e-01 0.0009637
#> [52,] 20  1.979e-01 0.0008781
#> [53,] 20  1.979e-01 0.0008001
#> [54,] 20  1.979e-01 0.0007290
#> [55,] 20  1.980e-01 0.0006643
#> [56,] 20  1.980e-01 0.0006052
#> [57,] 20  1.980e-01 0.0005515
#> [58,] 20  1.980e-01 0.0005025
#> [59,] 20  1.980e-01 0.0004578