Axing a rpart.

rpart objects are created from the rpart package, which is used for recursive partitioning for classification, regression and survival trees. This is where all the rpart specific documentation lies.

# S3 method for rpart
axe_call(x, verbose = FALSE, ...)

# S3 method for rpart
axe_ctrl(x, verbose = FALSE, ...)

# S3 method for rpart
axe_data(x, verbose = FALSE, ...)

# S3 method for rpart
axe_env(x, verbose = FALSE, ...)

Arguments

x
Model object.

verbose
Print information each time an axe method is executed that notes how much memory is released and what functions are disabled. Default is FALSE.

...
Any additional arguments related to axing.

Value

Axed model object.

Examples

# Load libraries
suppressWarnings(suppressMessages(library(parsnip)))
suppressWarnings(suppressMessages(library(rsample)))
suppressWarnings(suppressMessages(library(rpart)))

# Load data
set.seed(1234)
split <- initial_split(mtcars, props = 9/10)
car_train <- training(split)

# Create model and fit
rpart_fit <- decision_tree(mode = "regression") %>%
  set_engine("rpart") %>%
  fit(mpg ~ ., data = car_train, minsplit = 5, cp = 0.1)

butcher(rpart_fit)

#> n= 24 
#> 
#> node), split, n, deviance, yval
#>       * denotes terminal node
#> 
#> 1) root 24 732.7583 20.45833  
#>   2) hp>=101 17 121.2106 17.62353 *
#>   3) hp< 101 7 143.1571 27.34286 *