Axing a sclass object.

sclass objects are byproducts of classbagg objects. Since there are axe functions specific to this class, there is where all documentation related to sclass objects lies.

# S3 method for sclass
axe_call(x, verbose = FALSE, ...)

# S3 method for sclass
axe_env(x, verbose = FALSE, ...)

Arguments

x
sclass object.

verbose
Print information each time an axe method is executed that notes how much memory is released and what functions are disabled. Default is FALSE.

...
Any additional arguments related to axing.

Value

Axed sclass object.

Examples

# Load libraries
suppressWarnings(suppressMessages(library(ipred)))
suppressWarnings(suppressMessages(library(rpart)))
suppressWarnings(suppressMessages(library(MASS)))

# Load data
data("GlaucomaM", package = "TH.data")

classbagg_fit <- bagging(Class ~ ., data = GlaucomaM, coob = TRUE)

butcher(classbagg_fit$mtrees[[1]])

#> $bindx
#>   [1] 175 168 123  48 108 131  87  41 115  72 160  42  43   2 117  79 182  49
#>  [19] 102 184  51 134  49  49  57 136  26 145 191 123   8  96  22  35  41 131
#>  [37]  86  57  10 183 150 109 153 131  83 178  70  20 169 191  71  61  96 179
#>  [55]  70 127 155  72  48  60 193  36 119 156  19 137 126 158  63 116 145 102
#>  [73]  58 134  85  85 160  77  17 194 126   6 186 130 181  11  60 100 163  21
#>  [91] 170 102 134   9 171  19  22 174  40  29 109 185 144 194  38  88 158 107
#> [109]  70 147 119  79  86 142 151 188 147 147  71 123 188 176 123  99  26 144
#> [127]  22  28  98 101 176 149 169 163  40  67  72 147 189 121 196 120  84  83
#> [145]  96  17  25 163 130 101 175 175  30  10  22 185  49  89  80  73 148  12
#> [163]  37  81 195  47  86 166  82  94 147 146 151 107  31   6   7  72   4  43
#> [181] 189 143 145 171  39  42  35  58 146 139 184 136  22 170 101 116
#> 
#> $btree
#> n= 196 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>   1) root 196 92 1  
#>     2) varg< 0.209 92  8 1  
#>       4) as< 1.009 87  3 1  
#>         8) mhcg< 0.3135 86  2 1  
#>          16) vass< 0.016 82  0 1 *
#>          17) vass>=0.016 4  2 1  
#>            34) ag< 2.664 2  0 1 *
#>            35) ag>=2.664 2  0 2 *
#>         9) mhcg>=0.3135 1  0 2 *
#>       5) as>=1.009 5  0 2 *
#>     3) varg>=0.209 104 20 2  
#>       6) tmi>=0.0095 20  5 1  
#>        12) phcn>=-0.0895 15  0 1 *
#>        13) phcn< -0.0895 5  0 2 *
#>       7) tmi< 0.0095 84  5 2  
#>        14) mhcg>=0.165 2  0 1 *
#>        15) mhcg< 0.165 82  3 2  
#>          30) mhci>=0.2165 1  0 1 *
#>          31) mhci< 0.2165 81  2 2  
#>            62) at< 0.322 11  2 2  
#>             124) at>=0.3145 2  0 1 *
#>             125) at< 0.3145 9  0 2 *
#>            63) at>=0.322 70  0 2 *
#> 
#> attr(,"class")
#>                                 class 
#> "butchered_sclass"           "sclass"