Axing an survreg.

survreg objects are created from the survival package. They are returned from the survreg function, representing fitted parametric survival models.

# S3 method for survreg
axe_call(x, verbose = FALSE, ...)

# S3 method for survreg
axe_data(x, verbose = FALSE, ...)

# S3 method for survreg
axe_env(x, verbose = FALSE, ...)

Arguments

x
A model object.

verbose
Print information each time an axe method is executed. Notes how much memory is released and what functions are disabled. Default is FALSE.

...
Any additional arguments related to axing.

Value

Axed survreg object.

Examples

# Load libraries
suppressWarnings(suppressMessages(library(parsnip)))
suppressWarnings(suppressMessages(library(survival)))

# Create model and fit
survreg_fit <- surv_reg(mode = "regression", dist = "weibull") %>%
  set_engine("survival") %>%
  fit(Surv(futime, fustat) ~ 1, data = ovarian)

#> Warning: `surv_reg()` was deprecated in parsnip 0.1.6.
#> Please use `survival_reg()` instead.

out <- butcher(survreg_fit, verbose = TRUE)

#> ✔ Memory released: '2,136 B'

# Another survreg object
wrapped_survreg <- function() {
  some_junk_in_environment <- runif(1e6)
  fit <- survreg(Surv(time, status) ~ ph.ecog + age + strata(sex),
                 data = lung)
  return(fit)
}

# Remove junk
cleaned_survreg <- butcher(wrapped_survreg(), verbose = TRUE)

#> ✔ Memory released: '8,076,840 B'

#> ✖ Disabled: `print()`, `summary()`, `residuals()`

# Check size
lobstr::obj_size(cleaned_survreg)

#> 11,584 B