Axing an survreg.penal

survreg.penal objects are created from the survival package. They are returned from the survreg function, representing fitted parametric survival models.

Usage

# S3 method for survreg.penal
axe_call(x, verbose = FALSE, ...)

# S3 method for survreg.penal
axe_data(x, verbose = FALSE, ...)

# S3 method for survreg.penal
axe_env(x, verbose = FALSE, ...)

Arguments

x

A model object.

verbose

Print information each time an axe method is executed. Notes how much memory is released and what functions are disabled. Default is FALSE.

...

Any additional arguments related to axing.

Value

Axed survreg object.

Examples

# Load libraries
suppressWarnings(suppressMessages(library(parsnip)))
suppressWarnings(suppressMessages(library(survival)))
suppressWarnings(library(lobstr))

# Create model and fit
survreg_fit <- surv_reg(mode = "regression", dist = "weibull") %>%
  set_engine("survival") %>%
  fit(Surv(time, status) ~ rx, data = rats)
#> Warning: `surv_reg()` was deprecated in parsnip 0.1.6.
#> Please use `survival_reg()` instead.

out <- butcher(survreg_fit, verbose = TRUE)
#> ✔ Memory released: '7,392 B'

# Another survreg.penal object
wrapped_survreg.penal <- function() {
  some_junk_in_environment <- runif(1e6)
  fit <- survreg(Surv(time, status) ~ rx,
                 data = rats, subset = (sex == "f"))
  return(fit)
}

# Remove junk
cleaned_sp <- axe_env(wrapped_survreg.penal(), verbose = TRUE)
#> ✔ Memory released: '8,083,912 B'

# Check size
lobstr::obj_size(cleaned_sp)
#> 13,744 B