Axe data.

Remove the training data attached to modeling objects.

axe_data(x, verbose = FALSE, ...)

Arguments

x
A model object.

verbose
Print information each time an axe method is executed. Notes how much memory is released and what functions are disabled. Default is FALSE.

...
Any additional arguments related to axing.

Value

Model object without the training data

Methods

See the following help topics for more details about individual methods:butcher

axe-classbagg: classbagg

axe-earth: earth

axe-gausspr: gausspr

axe-ksvm: ksvm

axe-model_fit: model_fit

axe-rpart: rpart

axe-spark: ml_model

axe-survreg: survreg

axe-survreg.penal: survreg.penal

axe-train: train

axe-train.recipe: train.recipe