Axe an environment.

Remove the environment(s) attached to modeling objects as they are not required in the downstream analysis pipeline. If found, the environment is replaced with rlang::empty_env().

axe_env(x, verbose = FALSE, ...)

Arguments

x
A model object.

verbose
Print information each time an axe method is executed. Notes how much memory is released and what functions are disabled. Default is FALSE.

...
Any additional arguments related to axing.

Value

Model object with empty environments.

Methods

See the following help topics for more details about individual methods: butcher

axe-classbagg: classbagg

axe-flexsurvreg: flexsurvreg

axe-formula: formula

axe-function: function

axe-gausspr: gausspr

axe-kknn: kknn

axe-lm: lm

axe-mda: mda

axe-model_fit: model_fit

axe-nnet: nnet

axe-randomForest: randomForest

axe-recipe: quosure, recipe, step, step_arrange, step_bagimpute, step_bin2factor, step_BoxCox, step_bs, step_center, step_classdist, step_corr, step_count, step_date, step_depth, step_discretize, step_downsample, step_dummy, step_factor2string, step_filter, step_geodist, step_holiday, step_hyperbolic, step_ica, step_integer, step_interact, step_inverse, step_invlogit, step_isomap, step_knnimpute, step_kpca, step_lag, step_lincomb, step_log, step_logit, step_lowerimpute, step_meanimpute, step_medianimpute, step_modeimpute, step_mutate, step_naomit, step_nnmf, step_novel, step_ns, step_num2factor, step_nzv, step_ordinalscore, step_other, step_pca, step_pls, step_poly, step_range, step_ratio, step_regex, step_relu, step_rm, step_rollimpute, step_scale, step_shuffle, step_slice, step_spatialsign, step_sqrt, step_string2factor, step_unorder, step_upsample, step_window, step_YeoJohnson, step_zv

axe-rpart: rpart

axe-sclass: sclass

axe-stanreg: stanreg

axe-survreg: survreg

axe-survreg.penal: survreg.penal

axe-terms: terms

axe-train: train

axe-train.recipe: train.recipe

axe-xgb.Booster: xgb.Booster