Axe fitted values.

Remove the fitted values attached to modeling objects.

axe_fitted(x, verbose = FALSE, ...)

Arguments

x
A model object.

verbose
Print information each time an axe method is executed. Notes how much memory is released and what functions are disabled. Default is FALSE.

...
Any additional arguments related to axing.

Value

Model object without the fitted values.

Methods

See the following help topics for more details about individual methods:butcher

axe-C5.0: C5.0

axe-earth: earth

axe-gausspr: gausspr

axe-kknn: kknn

axe-ksvm: ksvm

axe-lm: lm

axe-mda: mda

axe-model_fit: model_fit

axe-nnet: nnet

axe-ranger: ranger

axe-spark: ml_model

axe-train: train

axe-train.recipe: train.recipe

axe-xgb.Booster: xgb.Booster