Axing a glmnet.

This is where all the glmnet specific documentation lies.

# S3 method for glmnet
axe_call(x, verbose = FALSE, ...)

Arguments

x
Model object.

verbose
Print information each time an axe method is executed that notes how much memory is released and what functions are disabled. Default is FALSE.

...
Any additional arguments related to axing.

Value

Axed model object.

Examples

library(glmnet)
x <- model.matrix(mpg ~ ., data = mtcars)
y <- as.matrix(sample(c(1, 0), size = 32, replace = TRUE))
fit <- glmnet(x, as.factor(y), family = "binomial")

butcher(fit)

#> 
#> Call:  dummy_call() 
#> 
#>       Df      %Dev    Lambda
#>  [1,]  0 5.176e-16 0.0974300
#>  [2,]  1 5.540e-03 0.0887800
#>  [3,]  1 1.012e-02 0.0808900
#>  [4,]  1 1.392e-02 0.0737000
#>  [5,]  1 1.707e-02 0.0671600
#>  [6,]  1 1.968e-02 0.0611900
#>  [7,]  1 2.185e-02 0.0557500
#>  [8,]  1 2.365e-02 0.0508000
#>  [9,]  1 2.514e-02 0.0462900
#> [10,]  1 2.638e-02 0.0421800
#> [11,]  1 2.741e-02 0.0384300
#> [12,]  1 2.827e-02 0.0350100
#> [13,]  2 2.957e-02 0.0319000
#> [14,]  2 3.211e-02 0.0290700
#> [15,]  2 3.424e-02 0.0264900
#> [16,]  2 3.602e-02 0.0241300
#> [17,]  2 3.751e-02 0.0219900
#> [18,]  3 3.946e-02 0.0200400
#> [19,]  4 4.884e-02 0.0182600
#> [20,]  5 6.202e-02 0.0166300
#> [21,]  6 7.364e-02 0.0151600
#> [22,]  6 8.373e-02 0.0138100
#> [23,]  7 9.397e-02 0.0125800
#> [24,]  7 1.045e-01 0.0114700
#> [25,]  8 1.150e-01 0.0104500
#> [26,]  8 1.246e-01 0.0095190
#> [27,]  8 1.331e-01 0.0086730
#> [28,]  8 1.408e-01 0.0079030
#> [29,]  8 1.477e-01 0.0072010
#> [30,]  9 1.540e-01 0.0065610
#> [31,]  9 1.601e-01 0.0059780
#> [32,]  9 1.657e-01 0.0054470
#> [33,]  9 1.707e-01 0.0049630
#> [34,]  9 1.751e-01 0.0045220
#> [35,]  9 1.790e-01 0.0041210
#> [36,] 10 1.835e-01 0.0037550
#> [37,] 10 1.874e-01 0.0034210
#> [38,] 10 1.909e-01 0.0031170
#> [39,] 10 1.940e-01 0.0028400
#> [40,] 10 1.966e-01 0.0025880
#> [41,] 10 1.989e-01 0.0023580
#> [42,] 10 2.008e-01 0.0021480
#> [43,] 10 2.025e-01 0.0019580
#> [44,] 10 2.039e-01 0.0017840
#> [45,] 10 2.052e-01 0.0016250
#> [46,] 10 2.062e-01 0.0014810
#> [47,] 10 2.071e-01 0.0013490
#> [48,] 10 2.078e-01 0.0012290
#> [49,] 10 2.085e-01 0.0011200
#> [50,] 10 2.090e-01 0.0010210
#> [51,] 10 2.095e-01 0.0009300
#> [52,] 10 2.099e-01 0.0008474
#> [53,] 10 2.102e-01 0.0007721
#> [54,] 10 2.105e-01 0.0007035
#> [55,] 10 2.107e-01 0.0006410
#> [56,] 10 2.109e-01 0.0005841
#> [57,] 10 2.111e-01 0.0005322
#> [58,] 10 2.112e-01 0.0004849
#> [59,] 10 2.113e-01 0.0004418
#> [60,] 10 2.114e-01 0.0004026
#> [61,] 10 2.115e-01 0.0003668
#> [62,] 10 2.116e-01 0.0003342
#> [63,] 10 2.116e-01 0.0003045
#> [64,] 10 2.117e-01 0.0002775
#> [65,] 10 2.117e-01 0.0002528
#> [66,] 10 2.118e-01 0.0002304
#> [67,] 10 2.118e-01 0.0002099
#> [68,] 10 2.118e-01 0.0001913
#> [69,] 10 2.118e-01 0.0001743
#> [70,] 10 2.119e-01 0.0001588
#> [71,] 10 2.119e-01 0.0001447
#> [72,] 10 2.119e-01 0.0001318
#> [73,] 10 2.119e-01 0.0001201