Axing a glmnet.

glmnet objects are created from the glmnet package, leveraged to fit generalized linear models via penalized maximum likelihood.

# S3 method for glmnet
axe_call(x, verbose = FALSE, ...)

Arguments

x
A model object.

verbose
Print information each time an axe method is executed. Notes how much memory is released and what functions are disabled. Default is FALSE.

...
Any additional arguments related to axing.

Value

Axed glmnet object.

Examples

library(glmnet)
x <- model.matrix(mpg ~ ., data = mtcars)
y <- as.matrix(sample(c(1, 0), size = 32, replace = TRUE))
fit <- glmnet(x, as.factor(y), family = "binomial")

butcher(fit)

#> 
#> Call:  dummy_call() 
#> 
#>       Df     %Dev    Lambda
#>  [1,]  0 0.000000 0.0761700
#>  [2,]  1 0.003669 0.0694000
#>  [3,]  1 0.006683 0.0632400
#>  [4,]  2 0.009301 0.0576200
#>  [5,]  2 0.012240 0.0525000
#>  [6,]  2 0.014680 0.0478400
#>  [7,]  2 0.016700 0.0435900
#>  [8,]  2 0.018390 0.0397200
#>  [9,]  2 0.019790 0.0361900
#> [10,]  2 0.020950 0.0329700
#> [11,]  3 0.025280 0.0300400
#> [12,]  3 0.032160 0.0273700
#> [13,]  4 0.039050 0.0249400
#> [14,]  5 0.046760 0.0227300
#> [15,]  5 0.061540 0.0207100
#> [16,]  7 0.077450 0.0188700
#> [17,]  7 0.096500 0.0171900
#> [18,]  7 0.112800 0.0156600
#> [19,]  7 0.126600 0.0142700
#> [20,]  7 0.138300 0.0130100
#> [21,]  8 0.149100 0.0118500
#> [22,]  8 0.160600 0.0108000
#> [23,]  9 0.171200 0.0098380
#> [24,]  9 0.181500 0.0089640
#> [25,]  9 0.190600 0.0081680
#> [26,]  8 0.198400 0.0074420
#> [27,]  8 0.205200 0.0067810
#> [28,]  8 0.211200 0.0061780
#> [29,]  8 0.216400 0.0056300
#> [30,]  8 0.221100 0.0051290
#> [31,]  8 0.225100 0.0046740
#> [32,]  8 0.228700 0.0042590
#> [33,]  8 0.231900 0.0038800
#> [34,]  8 0.234600 0.0035360
#> [35,]  8 0.237000 0.0032210
#> [36,]  8 0.239100 0.0029350
#> [37,]  8 0.240900 0.0026750
#> [38,]  8 0.242500 0.0024370
#> [39,]  8 0.243800 0.0022200
#> [40,]  8 0.245000 0.0020230
#> [41,]  8 0.246000 0.0018430
#> [42,]  8 0.246800 0.0016800
#> [43,]  8 0.247600 0.0015300
#> [44,]  8 0.248200 0.0013940
#> [45,]  8 0.248700 0.0012710
#> [46,]  9 0.249200 0.0011580
#> [47,]  9 0.249700 0.0010550
#> [48,]  9 0.250100 0.0009612
#> [49,]  9 0.250400 0.0008758
#> [50,]  9 0.250700 0.0007980
#> [51,]  9 0.250900 0.0007271
#> [52,] 10 0.251100 0.0006625
#> [53,] 10 0.251300 0.0006036
#> [54,] 10 0.251500 0.0005500
#> [55,] 10 0.251600 0.0005012
#> [56,] 10 0.251700 0.0004566
#> [57,] 10 0.251800 0.0004161
#> [58,] 10 0.251800 0.0003791
#> [59,] 10 0.251900 0.0003454
#> [60,] 10 0.251900 0.0003147
#> [61,] 10 0.252000 0.0002868
#> [62,] 10 0.252000 0.0002613
#> [63,] 10 0.252000 0.0002381
#> [64,] 10 0.252100 0.0002169
#> [65,] 10 0.252100 0.0001977
#> [66,] 10 0.252100 0.0001801
#> [67,] 10 0.252100 0.0001641
#> [68,] 10 0.252100 0.0001495
#> [69,] 10 0.252200 0.0001362
#> [70,] 10 0.252200 0.0001241