Reduce the Size of Modeling Objects

Reduce the size of modeling objects after fitting. These parsed-down versions of the original modeling object have been tested to work with their respective predict functions. Future iterations of this package should support additional analysis functions outside of just predict.

This package provides five S3 generics:

axe_call To remove the call object.

axe_ctrl To remove controls associated with training.

axe_data To remove the original data.

axe_env To remove inherited environments.

axe_fitted To remove fitted values.

These specific attributes of the model objects are chosen as they are often not required for downstream data analysis functions to work and are often the heaviest components of the fitted object. By calling the wrapper function butcher, all the sub-axe functions listed above are executed on the model object, returning a butchered model object that has an additional butcher class assignment. If only a specific axe_ function is called, the axed model object will also have the same addition of a butcher class assignment.

See also

Author

Maintainer: Max Kuhn max@posit.co (ORCID)

Authors:

Joyce Cahoon joyceyu48@gmail.com (ORCID)

Davis Vaughan davis@posit.co

Alex Hayes alexpghayes@gmail.com

Julia Silge julia.silge@posit.co (ORCID)

Other contributors:

Posit Software, PBC (ROR) [copyright holder, funder]